# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/match-ai": {
      "post": {
        "summary": "Match Users Ai",
        "operationId": "match_users_ai_match_ai_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/preview": {
      "post": {
        "summary": "Preview Smart Team",
        "operationId": "preview_smart_team_teams_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewSmartTeamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/confirm": {
      "post": {
        "summary": "Confirm Smart Team",
        "operationId": "confirm_smart_team_teams_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmSmartTeamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/teams/analyze": {
      "post": {
        "summary": "Analyze Team",
        "operationId": "analyze_team_teams_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeTeamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalyzeTeamRequest": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "avg_o": {
            "type": "number",
            "title": "Avg O"
          },
          "avg_c": {
            "type": "number",
            "title": "Avg C"
          },
          "avg_e": {
            "type": "number",
            "title": "Avg E"
          },
          "avg_a": {
            "type": "number",
            "title": "Avg A"
          },
          "avg_n": {
            "type": "number",
            "title": "Avg N"
          }
        },
        "type": "object",
        "required": [
          "score",
          "avg_o",
          "avg_c",
          "avg_e",
          "avg_a",
          "avg_n"
        ],
        "title": "AnalyzeTeamRequest"
      },
...
      "ConfirmSmartTeamRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "default": ""
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "title": "Deadline"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Start Date"
          },
          "leader_id": {
            "type": "string",
            "title": "Leader Id"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/SmartQuestRequirement"
            },
            "type": "array",
            "title": "Requirements"
          },
          "member_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Member Ids"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "title",
          "deadline",
          "start_date",
          "leader_id",
          "requirements",
          "member_ids",
          "status"
        ],
        "title": "ConfirmSmartTeamRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchRequest": {
        "properties": {
          "user1_id": {
            "type": "string",
            "title": "User1 Id"
          },
          "user2_id": {
            "type": "string",
            "title": "User2 Id"
          }
        },
        "type": "object",
        "required": [
          "user1_id",
          "user2_id"
        ],
        "title": "MatchRequest"
      },
...
      "PreviewSmartTeamRequest": {
        "properties": {
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/SmartQuestRequirement"
            },
            "type": "array",
            "title": "Requirements"
          },
          "candidate_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Candidate Ids"
          }
        },
        "type": "object",
        "required": [
          "requirements",
          "candidate_ids"
        ],
        "title": "PreviewSmartTeamRequest"
      },
...
      "SmartQuestRequirement": {
        "properties": {
          "department_id": {
            "type": "string",
            "title": "Department Id"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "department_id",
          "count"
        ],
        "title": "SmartQuestRequirement"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}